// frontend/src/api-generated/models/danhSachThietBiPhanMemResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface DanhSachThietBiPhanMemResponse {
  id?: number;
  idTaiSanPhanMem?: number;
  tenTaiSanPhanMem?: string;
  maMauTaiSanPhanMem?: string;
  idNhaCungCap?: number;
  idDonVi?: number;
  keyBanQuyen?: string;
  maChungTuMua?: string;
  tongSoGhe?: number;
  giaMua?: number;
  thoiGianMua?: string;
  thoiGianHetHan?: string;
  trangThaiKho?: string;
  trangThai?: string;
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
}